tithe-report/requirements.txt
streamlit>=1.20.0
pandas>=2.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
google-generativeai>=0.3.0
